refactor(react): Use @lit-labs/react
https://github.com/pocka/figspec/issues/34

The existing glue code was working okay. However, adding property or
attribute to the element is daunting and easy to misconfigure.
I replaced the code with new integration package provided by official
lit maintainers for maintanability reasons.

I kept type signatures to make sure this change does not break users'
type checking accidentally.

close #34
