--- packages/react/src/types.d.ts
@@ -1,6 +1,0 @@
-declare namespace JSX {
-  interface IntrinsicElements {
-    "figspec-frame-viewer": any;
-    "figspec-file-viewer": any;
-  }
-}
