+++ packages/react/tsconfig.json
@@ -3,11 +3,9 @@
     "target": "ES2015",
     "module": "ESNext",
     "declaration": true,
-    "jsx": "react",
-    "jsxFactory": "createElement",
     "lib": ["ESNext", "DOM"],
     "strict": true,
     "moduleResolution": "Node"
   },
-  "include": ["./src/**/*.ts", "./src/**/*.tsx"]
+  "include": ["./src/**/*.ts"]
 }
